\name{genes}
\alias{genes}
\docType{data}
\title{
    List of genes.
}
\description{
    Example of a list of genes that can be provided by an user.
}
\usage{data("genes")}
\format{
    A vector with 2 observations (ACAA1, SLC6A12).
}
\source{
    Sample of data from a study on chickens, under heat-stress condition.
}
\keyword{datasets}
