\name{meta}
\alias{meta}
\docType{data}
\title{
    List of metabolites.
}
\description{
    Example of a list of metabolites that can be provided by an user.
}
\usage{data("meta")}
\format{
    A vector with 2 observations (C00002, C00719).
}
\source{
    Sample of data from a study on chickens, under heat-stress condition.
}
\keyword{datasets}
