\name{MPINetData}
\alias{MPINetData}
\alias{getnodeseq}
\alias{getStr}
\alias{getBackground}
\alias{getpathList('KEGG')}
\alias{consensusPath}
\docType{data}

\title{The variables in the environment variable \code{MPINetData} of the system}
\description{
	  The variables in the environment variable \code{MPINetData} of the system.
}

\format{An environment variable}
\author{Yanjun Xu <tonghua605@163.com>, Chunquan Li <lcqbio@aliyun.com.cn> and Xia Li <lixia@hrbmu.edu.cn>}
\keyword{file}
